requirements.txt
alembic==1.14.1
annotated-types==0.7.0
asn1crypto==1.5.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
flask==3.1.0
flask-httpauth==4.8.0
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-secret-manager==2.23.0
googleapis-common-protos==1.67.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==21.2.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.5
mako==1.3.9
markupsafe==3.0.2
packaging==24.2
pg8000==1.31.2
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pyproject-hooks==1.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
requests==2.32.3
rsa==4.9
ruff==0.9.6
scramp==1.4.5
six==1.17.0
sqlalchemy==2.0.38
typing-extensions==4.12.2
urllib3==2.3.0
uuid-utils==0.10.0
uuid-v7==1.0.0
uuid7==0.1.0
werkzeug==3.1.3
